Extract canvas input controller
Split GameLoop so it composes the frame loop with a dedicated CanvasInputController for pointer, touch, wheel, drag-select, space-pan, and gesture state machines.

Keep the GameLoop public facade intact for existing callers and update engine exports/tests to use the new boundary.
