+++ engine/index.ts
@@ -14,11 +14,14 @@ export type {
 
 export { actionLayerController } from "./action-layer-controller.ts";
 
-export { GameLoop, gameLoop, SpacePanMode } from "./game-loop.ts";
-export type { CanvasRendererPort, InputState } from "./game-loop.ts";
+export { GameLoop, gameLoop } from "./game-loop.ts";
+export type { CanvasRendererPort } from "./frame-loop.ts";
 
-export { viewportAnimation } from "./viewport-animation.ts";
-export type { ViewportAnimationOptions } from "./viewport-animation.ts";
+export type { InputState, GameLoopDeps } from "./canvas-input-controller.ts";
+export { SpacePanMode } from "./canvas-input-controller.ts";
+
+export { viewportAnimation, ViewportAnimationController } from "./viewport-animation.ts";
+export type { ViewportAnimationOptions, ViewportStore } from "./viewport-animation.ts";
 
 export { disintegrationController } from "./disintegration-controller.ts";
 export { perfOverlay } from "./perf-overlay.ts";
